👷 Port CI to Github Actions

+++ scripts/update_docs.sh
@@ -21,9 +21,13 @@ cd "$(dirname "$(dirname "${SELF}")")"
 
 SITE_BASE="$(pwd)"
 
-# grab and unpack the latest matrix-docs build from buildkite
-rm -rf assets.tar.gz assets
-scripts/fetch-buildkite-artifact matrix-dot-org matrix-doc assets.tar.gz
+if ! [ -z ${BUILDKITE_ACCESS_TOKEN+x} ]; then
+  # grab and unpack the latest matrix-docs build from buildkite
+  rm -rf assets.tar.gz assets
+  scripts/fetch-buildkite-artifact matrix-dot-org matrix-doc assets.tar.gz
+fi
+
+# otherwise, assume the latest build was fetched by Github Actions.
 tar -xzf assets.tar.gz
 
 # copy the swagger UI into place
